edit ve removelar düzenlendi

+++ patient-management-front/src/Pages/Lists/HospitalList.js
@@ -15,6 +15,7 @@ const HospitalList = () => {
     mail: '',
     address: ''
   });
+  const [selectedHospitalId, setHospitalId] = useState(null);
   const [editMode, setEditMode] = useState(false);
 const [editHospitalId, setEditHospitalId] = useState(null);
   useEffect(() => {
@@ -80,34 +81,80 @@ const [editHospitalId, setEditHospitalId] = useState(null);
         console.error('An error occurred while adding admin:', error);
       }
     };
-  const handleRemoveHospital = async (hospitalId) => {
-    try {
-      const response = await fetch(`https://localhost:7050/RemoveHospital?hospitalId=${hospitalId}`, {
-        method: 'DELETE',
-      });
-      if (response.status === 200) {
-        console.log('Admin successfully removed');
-        fetchHospitals(); // Fetch updated admins after successful removal
-      } else {
-        console.error('Failed to remove admin');
+    const handleUpdate = async (e) => {
+      e.preventDefault();
+      
+      console.log(formData);
+      try {
+        const response = await fetch('https://localhost:7050/HospitalUpdate', {
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/json'
+          },
+          body: JSON.stringify({
+            Name: formData.hospitalName,
+            Phone: formData.phone,
+            Address: formData.address,
+            HospitalId: selectedHospitalId
+          })
+        });
+        if (response.ok) {
+          // Başarılı bir şekilde admin eklendiğinde yapılacak işlemler
+          console.log('Hospital successfully updated');
+          // Yeni bir veri çekme işlemi başlat
+          fetchHospitals();
+          // Form verilerini sıfırla
+          setFormData({
+            hospitalName: '',
+            phone: '',
+            mail: '',
+            address: ''
+          });
+          setEditMode(false);
+        } else {
+          console.error('Failed to update admin');
+        }
+      } catch (error) {
+        console.error('An error occurred while updating admin:', error);
       }
-    } catch (error) {
-      console.error('An error occurred while removing admin:', error);
+    };
+  const handleRemoveHospital = async (hospitalId) => {
+    const confirmation = window.confirm("Are you sure you want to remove this hospital?");
+    if(confirmation){
+      try {
+        const response = await fetch(`https://localhost:7050/RemoveHospital?hospitalId=${hospitalId}`, {
+          method: 'DELETE',
+        });
+        if (response.status === 200) {
+          console.log('Admin successfully removed');
+          fetchHospitals(); // Fetch updated admins after successful removal
+        } else {
+          console.error('Failed to remove admin');
+        }
+      } catch (error) {
+        console.error('An error occurred while removing admin:', error);
+      }
     }
+    
   };
   const handleEditHospital = (hospitalIndex) => {
-    const hospital = hospitals[hospitalIndex]; // İlgili hastanenin bilgilerini al
+    const hospital = hospitals[hospitalIndex]; 
     setFormData({
       hospitalName: hospital.name,
       phone: hospital.phone,
       address: hospital.address
     });
-    setEditMode(true); // Düzenleme modunu etkinleştir
-    toggleModal(); // Modalı aç
+    setEditMode(true); 
+    setHospitalId(hospital.hospitalId)
+    toggleModal2();
   };
   const toggleModal = () => {
     setShowModal(!showModal);
 
+  };
+  const toggleModal2 = () => {
+    setEditMode(!editMode);
+    
   };
   return (
     <body>
@@ -264,6 +311,38 @@ const [editHospitalId, setEditHospitalId] = useState(null);
                       </div>
                     </div>
                   )}
+                  {editMode && (
+                    <div class="xxx">
+                      <div class="popup-box">
+                        <form onSubmit={handleUpdate}>
+                        <div className="row mb-3">
+                            <div class="row mb-3">
+                              <label for="inputEmail3" class="col-sm-2 col-form-label">Hospital Name</label>
+                              <div class="col-sm-10">
+                                <input type="text" class="form-control" id="inputText" name="hospitalName" defaultValue={formData.hospitalName} onChange={handleInputChange} />
+                              </div>
+                            </div>
+                            <div class="row mb-3">
+                              <label for="inputEmail3" class="col-sm-2 col-form-label">Phone</label>
+                              <div class="col-sm-10">
+                                <input type="text" class="form-control" id="inputText" name="phone" defaultValue={formData.phone} onChange={handleInputChange} />
+                              </div>
+                            </div>
+                            <div class="row mb-3">
+                              <label for="inputEmail3" class="col-sm-2 col-form-label">Address</label>
+                              <div class="col-sm-10">
+                                <input type="text" class="form-control" id="inputText" name="address" defaultValue={formData.address} onChange={handleInputChange} />
+                              </div>
+                            </div>
+                          </div>
+                          <div className="text-center">
+                            <button type="submit" className="btn btn-primary">Update</button>
+                            <button className="btn-close-popup" onClick={toggleModal2}>Close</button>
+                          </div>
+                        </form>
+                      </div>
+                    </div>
+                  )}
                   <ul className="responsive-table">
                     <li className="table-header">
                       <div className="col col-2"><i className="ri ri-building-4-line"></i><span style={{ marginRight: '10px' }}></span>Hospital</div>
@@ -285,9 +364,6 @@ const [editHospitalId, setEditHospitalId] = useState(null);
                               <div className="col col-4" data-label="Mail">{hospital.phone}</div>
                               <div className="col col-5">
                                 <div className="btn-group">
-                                  <a className="btn btn-primary btn-sm" title="Open Admin" href=''>
-                                    <i class="bi bi-eye"></i>
-                                  </a>
                                   <span style={{ marginRight: 10 }} />
                                   <a className="btn btn-primary btn-sm" title="Edit Admin" onClick={() => handleEditHospital(index)}>
                                     <i className="ri ri-edit-2-fill"></i>
